pyproject.toml
[tool.poetry]
name = "advent-of-code"
version = "0.1.0"
description = ""
authors = ["JamesTheBard <jamesthebard@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-box = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
